refactor: split the 'apps/apps'code into 'config/apps' and 'config/config'
